// src/store/types/Nfe.ts
export interface Nfe {
	id: string;
	nameClient?: string;
	cpfClient?: string;
	addressClient?: string;
	districtClient?: string;
	zipCodeClient?: string;
	ufClient?: string;
	cityClient?: string;
	municipalRegistrationClient?: string;
	phoneClient?: string;
	nameCompany?: string;
	socialReasonCompany?: string;
	addressCompany?: string;
	zipCodeCompany?: string;
	cnpjCompany?: string;
	numberCompany?: string;
	itemInfo?: number;
	textInfo?: string;
	addressInfo?: string;
	valueInfo?: number;
	formOfPaymentPayment?: string;
	valuePayment?: string;
	maturityPayment?: string;
	valueTotal?: string;
	createdAt?: Date;
}

export interface NfeDeleteDto {
	id: string;
}